+++ restart-app.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
-set -x
-passenger-config restart-app /home/app/rails
+
+# restart-app.sh
+# This script is part of dora -- Docker container for Rails
+# https://github.com/bovender/dora
+
+APP_DIR=/home/app/rails
+echo "# dora restarting Rails app in $APP_DIR..."
+passenger-config restart-app $APP_DIR