+++ set-timezone.sh
@@ -1,6 +1,10 @@
 #!/usr/bin/env bash
 
+# set-timezone.sh
+# This script is part of dora -- Docker container for Rails
+# https://github.com/bovender/dora
+
 if [ "$TIMEZONE" != "" ]; then
-  echo "Setting time zone to $TIMEZONE"
+  echo "# dora setting time zone to $TIMEZONE"
   ln -sf "/usr/share/zoneinfo/$TIMEZONE" /etc/localtime
 fi