+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixed
 
 - Use `passenger-ruby27` rather than `passenger-ruby26` as base image.
+- Copy the right `rails-console.sh` helper script into the image.
 
 ## Version 2.0.1 (2020-09-13)